Initial Commit: Open-source the dotnet Hyper API
Since the initial release of [Hyper API](https://tableau.github.io/hyper-db/)
in 2019, Hyper API shipped for Python, Java, C++ and .Net. Usage of .Net never
really picked up and we were not able to identify any users using the .Net
version in production, so the Hyper team decided in September 2023 to deprecate
the .Net version of Hyper API. We are releasing the source code to the
open-source community, such that any interested .Net enthusiasts have an
opportunity to pick up where we left.